+++ src/core/reflection/sobject/reflect-custom-object-sources.ts
@@ -8,6 +8,7 @@ import { pipe } from 'fp-ts/function';
 import * as A from 'fp-ts/Array';
 import * as E from 'fp-ts/Either';
 import { CustomFieldMetadata } from './reflect-custom-field-source';
+import { getPickListValues } from './parse-picklist-values';
 
 export type CustomObjectMetadata = {
   type_name: 'customobject';
@@ -38,6 +39,7 @@ function reflectCustomObjectSource(
     E.flatMap(validate),
     E.map(toObjectMetadata),
     E.map((metadata) => addName(metadata, objectSource.name)),
+    E.map(parseInlineFields),
     E.map(addTypeName),
     E.map((metadata) => toParsedFile(objectSource.filePath, metadata)),
     E.mapLeft((error) => new ReflectionErrors([new ReflectionError(objectSource.filePath, error.message)])),
@@ -75,6 +77,42 @@ function addName(objectMetadata: CustomObjectMetadata, name: string): CustomObje
   return {
     ...objectMetadata,
     name,
+  };
+}
+
+function parseInlineFields(metadata: CustomObjectMetadata): CustomObjectMetadata {
+  // if "fields" is present, it might be a single object (if it only has one field)
+  // or an array
+  if (!Array.isArray(metadata.fields)) {
+    metadata.fields = [metadata.fields];
+  }
+
+  return {
+    ...metadata,
+    fields: metadata.fields.map((field) => convertInlineFieldsToCustomFieldMetadata(field, metadata.name)),
+  };
+}
+
+function convertInlineFieldsToCustomFieldMetadata(
+  inlineField: Record<string, unknown>,
+  parentName: string,
+): CustomFieldMetadata {
+  // Based on Salesforce's documentation, the only required field is "fullName"
+  const name = inlineField.fullName as string;
+  const description = inlineField.description ? (inlineField.description as string) : null;
+  const label = inlineField.label ? (inlineField.label as string) : name;
+  const type = inlineField.type ? (inlineField.type as string) : null;
+  const required = inlineField.required ? (inlineField.required as boolean) : false;
+
+  return {
+    type_name: 'customfield',
+    description,
+    label,
+    name,
+    parentName,
+    type,
+    required,
+    pickListValues: getPickListValues(inlineField),
   };
 }