+++ src/core/renderables/types.d.ts
@@ -189,11 +189,12 @@ export type RenderableCustomField = {
   heading: string;
   apiName: string;
   description: RenderableContent[];
-  pickListValues?: RenderableSection<string[]>
+  pickListValues?: RenderableSection<string[]>;
   type: 'field';
   fieldType?: string | null;
+  required: boolean;
 };
 
 export type Renderable = (RenderableClass | RenderableInterface | RenderableEnum | RenderableCustomObject) & {
-  filePath: string;
+  filePath: string | undefined;
 };